move exit queue position to service
