+++ src/containers/transactions/Exits.js
@@ -20,8 +20,6 @@ import { useSelector } from 'react-redux';
 import truncate from 'truncate-middle';
 
 import config from 'util/config';
-import networkService from 'services/networkService';
-import { selectAllQueues, selectQueues } from 'selectors/queueSelector';
 import { selectPendingExits, selectExitedExits } from 'selectors/exitSelector';
 
 import ProcessExitsModal from 'containers/modals/processexit/ProcessExitsModal';
@@ -36,41 +34,10 @@ function Exits ({ searchHistory }) {
   const [ page, setPage ] = useState(1);
   const [ processExitModal, setProcessExitModal ] = useState(false);
 
-  const queues = useSelector(selectAllQueues, isEqual);
-  const rawQueues = useSelector(selectQueues, isEqual);
   const pendingExits = orderBy(useSelector(selectPendingExits, isEqual), i => i.blockNumber, 'desc');
   const exitedExits = orderBy(useSelector(selectExitedExits, isEqual), i => i.blockNumber, 'desc');
 
-  // add extra data to pending exits using data from the queue
-  function enhanceExits (exits) {
-    return exits.map(i => {
-      if (i.returnValues) {
-        const exitId = networkService.web3.utils.hexToNumberString(i.returnValues.exitId._hex);
-        const queuedExit = queues.find(i => i.exitId === exitId);
-        let queuePosition;
-        let queueLength;
-        if (queuedExit) {
-          const tokenQueue = rawQueues[queuedExit.currency];
-          queuePosition = tokenQueue.findIndex(x => x.exitId === exitId);
-          queueLength = tokenQueue.length;
-        }
-        return {
-          ...i,
-          ...queuedExit
-            ? {
-              exitableAt: queuedExit.exitableAt,
-              currency: queuedExit.currency,
-              queuePosition: queuePosition + 1,
-              queueLength
-            }
-            : {}
-        };
-      }
-      return i;
-    });
-  }
-
-  const _pendingExits = enhanceExits(pendingExits).filter(i => {
+  const _pendingExits = pendingExits.filter(i => {
     return i.transactionHash.includes(searchHistory);
   });
 
@@ -81,11 +48,22 @@ function Exits ({ searchHistory }) {
   const renderPending = _pendingExits.map((i, index) => {
     const exitableMoment = moment.unix(i.exitableAt);
     const isExitable = moment().isAfter(exitableMoment);
+
+    function getStatus () {
+      if (!i.exitableAt) {
+        return 'Loading Status';
+      }
+      if (i.status === 'Confirmed' && i.pendingPercentage >= 100) {
+        return 'In Challenge Period';
+      }
+      return i.status;
+    }
+
     return (
       <Transaction
         key={`pending-${index}`}
         button={
-          isExitable
+          i.exitableAt && isExitable
             ? {
               onClick: () => setProcessExitModal(i),
               text: 'Process Exit'
@@ -93,11 +71,7 @@ function Exits ({ searchHistory }) {
             : undefined
         }
         link={`${config.etherscanUrl}/tx/${i.transactionHash}`}
-        status={
-          i.status === 'Confirmed' && i.pendingPercentage >= 100
-            ? 'In Challenge Period'
-            : i.status
-        }
+        status={getStatus()}
         subStatus={`Block ${i.blockNumber}`}
         statusPercentage={i.pendingPercentage <= 100 ? i.pendingPercentage : ''}
         title={truncate(i.transactionHash, 10, 4, '...')}
